move client package to utils
