+++ pkg/utils/httpclient/round_trippper.go
@@ -1,4 +1,4 @@
-package client
+package httpclient
 
 import (
 	"fmt"
