+++ pkg/utils/httpclient/round_trippper_test.go
@@ -1,4 +1,4 @@
-package client
+package httpclient
 
 import (
 	"net/http"
